refactor: Rename +state to data-access:
- Update imports for data-access folder
- Update config files for data-access folder
