+++ libs/data-access/jest.config.ts
@@ -3,7 +3,7 @@ export default {
    displayName: 'state',
    preset: '../../jest.preset.js',
    setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
-   coverageDirectory: '../../coverage/libs/+state',
+   coverageDirectory: '../../coverage/libs/data-access',
    transform: {
       '^.+\\.(ts|mjs|js|html)$': [
          'jest-preset-angular',
