+++ libs/data-access/project.json
@@ -1,7 +1,7 @@
 {
    "name": "state",
    "$schema": "../../node_modules/nx/schemas/project-schema.json",
-   "sourceRoot": "libs/+state/src",
+   "sourceRoot": "libs/data-access/src",
    "prefix": "state",
    "tags": [],
    "projectType": "library",
@@ -10,7 +10,7 @@
          "executor": "@nx/jest:jest",
          "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
          "options": {
-            "jestConfig": "libs/+state/jest.config.ts"
+            "jestConfig": "libs/data-access/jest.config.ts"
          }
       },
       "lint": {
