+++ libs/data-access/src/lib/adapters/index.ts
