+++ libs/data-access/src/lib/adapters/shared/index.ts
