+++ libs/data-access/src/lib/adapters/shared/ui.adapter.ts
