+++ libs/data-access/src/lib/adapters/train-arrivals/train.ui.adapter.ts
