+++ libs/data-access/src/lib/models/state.interfaces.ts
