+++ src/app/app.config.ts
@@ -1,7 +1,9 @@
 import { ApplicationConfig, importProvidersFrom, isDevMode } from '@angular/core';
 import { provideRouter, withEnabledBlockingInitialNavigation } from '@angular/router';
 import { appRoutes } from './app.routes';
-import { StateModule } from '@atl-transit/global-state';
+// TODO: Edit config for scoped packages and libraries
+// eslint-disable-next-line @nx/enforce-module-boundaries
+import { StateModule } from '@atl-transit/data-access';
 import { PathLocationStrategy, LocationStrategy } from '@angular/common';
 import { provideServiceWorker } from '@angular/service-worker';
