+++ tsconfig.base.json
@@ -16,7 +16,7 @@
       "baseUrl": ".",
       "paths": {
          "@atl-transit/features-ui": ["libs/features/src/index.ts"],
-         "@atl-transit/global-state": ["libs/+state/src/index.ts"],
+         "@atl-transit/data-access": ["libs/data-access/src/index.ts"],
          "@atl-transit/navigation": ["libs/features/src/navigation/index.ts"],
          "@atl-transit/search": ["libs/features/src/search/index.ts"],
          "@atl-transit/shared": ["libs/shared/src/index.ts"],
